♻️ refactor : error_dialogを配置するディレクトリを変更
