+++ lib/util/extension/widget_ref_extension.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../feature/error/presentation/error_dialog.dart';
+import '../../common_widget/error_dialog.dart';
 import '../../feature/loading/application/loading_state.dart';
 
 /// ダイアログ表示用のGlobalKeyを返す
